Refactoring package Domain
