+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/domain/MoreDetailsDomain.kt
@@ -1,5 +1,19 @@
 package ayds.lisboa.songinfo.moredetails.fulllogic.domain
+import ayds.lisboa.songinfo.moredetails.fulllogic.ArtistData
+import ayds.lisboa.songinfo.moredetails.fulllogic.domain.repositoryInterface.ArtistRepository
+import ayds.observer.Observable
+import ayds.observer.Subject
 
-interface MoreDetailsDomain {}
+interface MoreDetailsDomain {
+    val moreInformationObservable: Observable<ArtistData>
+    fun getArtistMoreInformation(term: String)
+}
 
-internal class MoreDetailsDomainImpl {}
+internal class MoreDetailsDomainImpl(private val repository: ArtistRepository): MoreDetailsDomain {
+    override val moreInformationObservable = Subject<ArtistData>()
+    override fun getArtistMoreInformation(artistName: String) {
+        repository.getArtistData(artistName).let {
+            moreInformationObservable.notify(it)
+        }
+    }
+}
