+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/domain/entities/ArtistData.kt
@@ -1,3 +1,8 @@
 package ayds.lisboa.songinfo.moredetails.fulllogic.domain.entities
 
-class ArtistData {}
+data class ArtistData(
+    var artistName: String?,
+    var infoArtist: String?,
+    var url: String = "",
+    var isLocallyStored: Boolean = false
+)
