+++ app/src/main/java/ayds/lisboa/songinfo/moredetails/fulllogic/domain/repositoryInterface/ArtistRepository.kt
@@ -1,0 +1,6 @@
+package ayds.lisboa.songinfo.moredetails.fulllogic.domain.repositoryInterface;
+import ayds.lisboa.songinfo.moredetails.fulllogic.ArtistData;
+
+public interface ArtistRepository {
+    fun getArtistData(artistName: String) : ArtistData
+}
